Implementing code
We are adding a `code` field to the profiling. Its meant to be a unique (per repo) string that can identify which profiling is being made on this system without needing to pass env, version or ids on the individual profiling uploads

Also adding a `needs_version_creation` in case the customer doesn't want that API call to be made on startup (they will need to do at least themselves once). The call itself is idempotent, so they should not need to worry about it unless they particularly need it to go